requirements.txt
attrs==25.3.0
aws-cdk-lib==2.199.0
cdk-nag==2.36.26
aws-cdk.asset-awscli-v1==2.2.237
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.cloud-assembly-schema==41.2.0
cattrs==24.1.3
constructs==10.4.2
exceptiongroup==1.3.0
importlib_resources==6.5.2
iniconfig==2.1.0
jsii==1.112.0
packaging==25.0
pluggy==1.6.0
publication==0.0.3
Pygments==2.19.1
pytest==8.4.0
python-dateutil==2.9.0.post0
six==1.17.0
tomli==2.2.1
typeguard==2.13.3
typing_extensions==4.14.0
zipp==3.23.0
